requirements.txt
datasets>=2.20.0
matplotlib==3.10.0
numpy==2.2.1
pandas==2.2.3
Pillow==11.1.0
pydicom>=3.0.0
prefetch_generator==1.0.3
pynvml==12.0.0
PyYAML==6.0.2
rich==13.9.4
scikit_learn==1.6.1
torch==2.5.1
torchvision==0.20.1
